pyproject.toml
[project]
name = "agents-playground"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=7.1.0",
    "jupyter>=1.1.1",
    "langchain>=1.1.0",
    "langchain-anthropic>=1.2.0",
    "langchain-core>=1.1.0",
    "langchain-mcp-adapters>=0.1.0",
    "langgraph>=1.0.4",
    "prompt-toolkit>=3.0.0",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
]

[tool.uv]
package = true

[tool.setuptools]
packages = ["agentchat"]

[project.scripts]
agentchat-direct = "agentchat:main_direct"
agentchat-programmatic = "agentchat:main_programmatic"

[dependency-groups]
dev = [
    "mypy>=1.19.0",
    "nbqa>=1.9.1",
    "pre-commit>=4.5.0",
    "ruff>=0.14.7",
]

[tool.ruff]
target-version = "py313"
line-length = 120
exclude = ["tmp", "nbconvert_templates"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "UP",     # pyupgrade
]
ignore = []

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_ignores = true
exclude = ["tmp", "nbconvert_templates"]
plugins = ["pydantic.mypy"]
disable_error_code = ["top-level-await"]

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
